Edit open end loop
