+++ packages/tools/src/utilities/math/polyline/index.ts
@@ -2,5 +2,16 @@ import {
   getFirstIntersectionWithPolyline,
   getClosestIntersectionWithPolyline,
 } from './getIntersectionWithPolyline';
+import getSpacingAndXYDirections from './getSpacingAndXYDirections';
+import pointsAreWithinCloseContourProximity from './pointsAreWithinCloseContourProximity';
+import addCanvasPointsToArray from './addCanvasPointsToArray';
+import pointCanProjectOnLine from './pointCanProjectOnLine';
 
-export { getFirstIntersectionWithPolyline, getClosestIntersectionWithPolyline };
+export {
+  getFirstIntersectionWithPolyline,
+  getClosestIntersectionWithPolyline,
+  getSpacingAndXYDirections,
+  pointsAreWithinCloseContourProximity,
+  addCanvasPointsToArray,
+  pointCanProjectOnLine,
+};
