+++ packages/tools/src/utilities/math/polyline/pointsAreWithinCloseContourProximity.ts
@@ -1,0 +1,11 @@
+import { vec2 } from 'gl-matrix';
+
+const pointsAreWithinCloseContourProximity = (
+  point1,
+  point2,
+  closeContourProximity
+): boolean => {
+  return vec2.dist(point1, point2) < closeContourProximity;
+};
+
+export default pointsAreWithinCloseContourProximity;
